+++ android/samples/test-app/src/main/java/com/liferay/mobile/screens/testapp/MainActivity.java
@@ -50,6 +50,7 @@ public class MainActivity extends ThemeActivity implements View.OnClickListener 
 		findViewById(R.id.user_display).setOnClickListener(this);
 		findViewById(R.id.image_display).setOnClickListener(this);
 		findViewById(R.id.gallery).setOnClickListener(this);
+		findViewById(R.id.portlet_display).setOnClickListener(this);
 	}
 
 	@Override
@@ -161,6 +162,11 @@ public class MainActivity extends ThemeActivity implements View.OnClickListener 
 				break;
 			case R.id.image_display:
 				start(ImageDisplayActivity.class);
+				break;
+			case R.id.portlet_display:
+				Intent intentPortlet = getIntentWithTheme(PortletDisplayActivity.class);
+				intentPortlet.putExtra("url", getResources().getString(R.string.liferay_portlet_url));
+				DefaultAnimation.startActivityWithAnimation(this, intentPortlet);
 				break;
 			default:
 				start(LoginActivity.class);